requirements.txt
asgiref==3.4.1
asn1crypto==1.4.0
cffi==1.14.6
cryptography==3.4.7
Django==3.2.5
django-ca==1.18.0
django-extensions==3.1.3
django-object-actions==3.0.2
idna==3.2
packaging==21.0
pycparser==2.20
pycryptodome==3.10.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
Werkzeug==2.0.1
